requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==23.1.0
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
clr-loader==0.2.6
cryptography==41.0.4
cssselect2==0.7.0
exceptiongroup==1.1.3
fonttools==4.42.1
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
gspread==5.7.2
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.26.0
oauthlib==3.2.2
oscrypto==1.3.0
outcome==1.2.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pyparsing==3.1.1
pypdf==3.16.1
pyphen==0.14.0
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pythonnet==3.0.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
svglib==1.5.1
tinycss2==1.2.1
trio==0.22.2
trio-websocket==0.10.4
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.16
weasyprint==60.0
webdriver-manager==4.0.0
webencodings==0.5.1
wsproto==1.2.0
zopfli==0.2.3
